Extract `NumericBoundParser` class
