+++ model/src/main/kotlin/io/spine/validation/bound/KotlinNumericBound.kt
@@ -27,26 +27,6 @@
 package io.spine.validation.bound
 
 import io.spine.base.FieldPath
-import io.spine.base.fieldPath
-import io.spine.protodata.Compilation
-import io.spine.protodata.ast.PrimitiveType
-import io.spine.protodata.ast.PrimitiveType.TYPE_DOUBLE
-import io.spine.protodata.ast.PrimitiveType.TYPE_FIXED32
-import io.spine.protodata.ast.PrimitiveType.TYPE_FIXED64
-import io.spine.protodata.ast.PrimitiveType.TYPE_FLOAT
-import io.spine.protodata.ast.PrimitiveType.TYPE_INT32
-import io.spine.protodata.ast.PrimitiveType.TYPE_INT64
-import io.spine.protodata.ast.PrimitiveType.TYPE_SFIXED32
-import io.spine.protodata.ast.PrimitiveType.TYPE_SFIXED64
-import io.spine.protodata.ast.PrimitiveType.TYPE_SINT32
-import io.spine.protodata.ast.PrimitiveType.TYPE_SINT64
-import io.spine.protodata.ast.PrimitiveType.TYPE_UINT32
-import io.spine.protodata.ast.PrimitiveType.TYPE_UINT64
-import io.spine.protodata.ast.name
-import io.spine.protodata.ast.qualifiedName
-import io.spine.protodata.check
-import io.spine.protodata.type.resolve
-import io.spine.validation.bound.BoundFieldSupport.numericPrimitives
 
 /**
  * One-to-one Kotlin representation of [NumericBound].
@@ -119,140 +99,3 @@ internal fun KotlinNumericBound.toProto(): NumericBound {
     }
     return builder.build()
 }
-
-/**
- * Parses the given raw [bound] string to a [KotlinNumericBound].
- *
- * For number-based bounds, the method checks the following:
- *
- * 1) The bound value is not empty.
- * 2) The provided number has `.` for floating-point fields, and does not have `.`
- *    for integer fields.
- * 3) The provided number fits into the range of the target field type.
- *
- * For field-based bounds:
- *
- * 1) The specified field path points to an existing field.
- * 2) The field bound is not referencing the field it restricts (self-referencing).
- * 3) The referenced field is of singular numeric type.
- *
- * Any violation of the above conditions leads to a compilation error.
- *
- * @return The parsed numeric bound.
- */
-internal fun NumericOptionMetadata.checkNumericBound(
-    bound: String,
-    exclusive: Boolean
-): KotlinNumericBound {
-    Compilation.check(bound.isNotEmpty(), file, field.span) {
-        "The `($optionName)` option could not parse the bound value specified for" +
-                " `${field.qualifiedName}` field because it is empty. Please provide either" +
-                " a numeric value or a field reference."
-    }
-    return if (bound.isFieldReference()) {
-        checkFieldValue(bound, exclusive)
-    } else {
-        checkNumberValue(bound, exclusive)
-    }
-}
-
-private fun NumericOptionMetadata.checkNumberValue(
-    number: String,
-    exclusive: Boolean
-): KotlinNumericBound {
-    if (fieldType in listOf(TYPE_FLOAT, TYPE_DOUBLE)) {
-        Compilation.check(FLOAT.matches(number), file, field.span) {
-            "The `($optionName)` option could not parse the `$number` bound value specified for" +
-                    " `${field.qualifiedName}` field. Please make sure the provided value is" +
-                    " a floating-point number. Examples: `12.3`, `-0.1`, `6.02E2`."
-        }
-    } else {
-        Compilation.check(INTEGER.matches(number), file, field.span) {
-            "The `($optionName)` option could not parse the `$number` bound value specified for" +
-                    " `${field.qualifiedName}` field. Please make sure the provided value is" +
-                    " an integer number. Examples: `123`, `-567823`."
-        }
-    }
-
-    val parsed = when (fieldType) {
-        TYPE_FLOAT -> number.toFiniteFloatOrNull()
-        TYPE_DOUBLE -> number.toFiniteDoubleOrNull()
-        TYPE_INT32, TYPE_SINT32, TYPE_SFIXED32 -> number.toIntOrNull()
-        TYPE_INT64, TYPE_SINT64, TYPE_SFIXED64 -> number.toLongOrNull()
-        TYPE_UINT32, TYPE_FIXED32 -> number.toUIntOrNull()
-        TYPE_UINT64, TYPE_FIXED64 -> number.toULongOrNull()
-        else -> unexpectedPrimitiveType(fieldType)
-    }
-
-    Compilation.check(parsed != null, file, field.span) {
-        "The `($optionName)` option could not parse the `$number` bound value specified for" +
-                " `${field.qualifiedName}` field. The value is out of range for the field" +
-                " type `${field.type.name}` the option is applied to."
-    }
-
-    return KotlinNumericBound(parsed!!, exclusive)
-}
-
-private fun NumericOptionMetadata.checkFieldValue(
-    fieldPath: String,
-    exclusive: Boolean
-): KotlinNumericBound {
-    Compilation.check(fieldPath != field.name.value, file, field.span) {
-        "The `($optionName)` option cannot use `$fieldPath` field as a bound value for" +
-                " the `${field.qualifiedName}` because self-referencing is prohibited." +
-                " Please use other message fields."
-    }
-
-    val boundFieldPath = fieldPath {
-        fieldName.addAll(fieldPath.split("."))
-    }
-
-    val boundField = try {
-        val (messageType, _) = typeSystem.findMessage(field.declaringType)!!
-        typeSystem.resolve(boundFieldPath, messageType)
-    } catch (e: IllegalStateException) {
-        Compilation.error(file, field.span) {
-            "The `($optionName)` option could not parse the `$fieldPath` field path specified" +
-                    " for `${field.qualifiedName}` field. Please make sure the provided field" +
-                    " path is valid: `${e.message}`."
-        }
-    }
-
-    val boundFieldType = boundField.type.primitive
-    Compilation.check(boundFieldType in numericPrimitives, file, field.span) {
-        "The `($optionName)` option cannot use `$fieldPath` field as a bound value for" +
-                " the `${field.qualifiedName}` field due to its type `${boundFieldType.name}`." +
-                " Only singular numeric fields are supported."
-    }
-
-    return KotlinNumericBound(boundFieldPath, exclusive)
-}
-
-/**
- * Parses this [String] to a [Float], returning null if the string is not a valid finite float.
- */
-private fun String.toFiniteFloatOrNull(): Float? =
-    toFloatOrNull()?.takeIf { it.isFinite() }
-
-/**
- * Parses this [String] to a [Double], returning null if the string is not a valid finite double.
- */
-private fun String.toFiniteDoubleOrNull(): Double? =
-    toDoubleOrNull()?.takeIf { it.isFinite() }
-
-
-/**
- * Tells whether this [String] bound value contains a field reference.
- */
-private fun String.isFieldReference() = first().run {
-    isLetter() || this == '_'
-}
-
-private fun unexpectedPrimitiveType(primitiveType: PrimitiveType): Nothing =
-    error(
-        "`KotlinNumericBound` cannot be created for `$primitiveType` field type." +
-                " Please make sure the policy correctly filtered unsupported field types."
-    )
-
-private val INTEGER = Regex("[-+]?\\d+")
-private val FLOAT = Regex("[-+]?\\d+\\.\\d+([eE][-+]?\\d+)?")
